backend/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-resized==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
gunicorn==23.0.0
packaging==25.0
Pillow==9.2.0
PyJWT==2.6.0
pytz==2022.5
setuptools==80.9.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.5
whitenoise==6.9.0
channels==3.0.5
channels-redis==4.1.0
daphne==3.0.2
